Added error checks for all end points + fixed a bug with reservation on stock

+++ src/books/services/books.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { createBookDto } from 'src/books/dtos/createBook.dto';
 import { dtoToEntity } from 'src/books/mapper/book.mapper';
@@ -16,13 +16,11 @@ import { BookStock } from 'src/typeorm/entities/BookStock';
 export class BooksService {
   constructor(
     @InjectRepository(Book) private booksRepository: Repository<Book>,
-    @InjectRepository(Publisher)
-    private publishersRepository: Repository<Publisher>,
+    @InjectRepository(Publisher) private publishersRepository: Repository<Publisher>,
     @InjectRepository(Author) private authorRepository: Repository<Author>,
     @InjectRepository(Genre) private genreRepository: Repository<Genre>,
-    @InjectRepository(Audit) private loggerRepo: Repository<Audit>,
-    @InjectRepository(Reservation)
-    private reservationRepo: Repository<Reservation>,
+    @InjectRepository(Audit) private loggerRepository: Repository<Audit>,
+    @InjectRepository(Reservation) private reservationRepo: Repository<Reservation>,
     @InjectRepository(BookStock) private bookStockRepo: Repository<BookStock>,
   ) {}
 
@@ -31,66 +29,69 @@ export class BooksService {
   }
 
   async createBook(bookDetails: createBookDto) {
-    const publishers = await this.publishersRepository.findByIds(
-      bookDetails.publisherIds,
-    );
-    const authors = await this.authorRepository.findByIds(
-      bookDetails.authorIds,
-    );
+    const publishers = await this.publishersRepository.findByIds(bookDetails.publisherIds);
+    const authors = await this.authorRepository.findByIds(bookDetails.authorIds);
     const genres = await this.genreRepository.findByIds(bookDetails.genreIds);
+
+    if (!publishers.length || !authors.length || !genres.length) {
+      throw new NotFoundException('Publishers, Authors, or Genres not found');
+    }
 
     let newBook: Book = dtoToEntity(bookDetails);
     newBook.addPublishers(publishers);
     newBook.addAuthors(authors);
     newBook.addGenres(genres);
-    let newLog: Audit = entityToLog('New Book', newBook, 'Books');
-    this.loggerRepo.save(newLog);
-    return this.booksRepository.save(newBook);
+
+    const newLog: Audit = entityToLog('New Book', newBook, 'Books');
+    await this.loggerRepository.save(newLog);
+
+    return await this.booksRepository.save(newBook);
   }
 
   async showByAtoZ() {
-    return this.booksRepository.find({
+    return await this.booksRepository.find({
       order: {
         bookTitle: 'ASC',
       },
     });
   }
+
   async showByZtoA() {
-    return this.booksRepository.find({
+    return await this.booksRepository.find({
       order: {
         bookTitle: 'DESC',
       },
     });
   }
+
   async showByGenre() {
-    return this.booksRepository
+    return await this.booksRepository
       .createQueryBuilder()
       .select('subject')
-      .addSelect(
-        "JSON_ARRAYAGG(JSON_OBJECT('bookTitle', bookTitle, 'id', id)) as books",
-      )
+      .addSelect("JSON_ARRAYAGG(JSON_OBJECT('bookTitle', bookTitle, 'id', id)) as books")
       .groupBy('subject')
       .getRawMany();
   }
+
   async showByCopyYear() {
-    return this.booksRepository
+    return await this.booksRepository
       .createQueryBuilder()
       .select('copyWriteYear')
-      .addSelect(
-        "JSON_ARRAYAGG(JSON_OBJECT('bookTitle', bookTitle, 'id', id)) as books",
-      )
+      .addSelect("JSON_ARRAYAGG(JSON_OBJECT('bookTitle', bookTitle, 'id', id)) as books")
       .groupBy('copyWriteYear')
       .orderBy('copyWriteYear', 'DESC')
       .getRawMany();
   }
+
   async updateBook(id: number, bookDetails: createBookDto) {
-    const publishers = await this.publishersRepository.findByIds(
-      bookDetails.publisherIds,
-    );
-    const authors = await this.authorRepository.findByIds(
-      bookDetails.authorIds,
-    );
+    const publishers = await this.publishersRepository.findByIds(bookDetails.publisherIds);
+    const authors = await this.authorRepository.findByIds(bookDetails.authorIds);
     const genres = await this.genreRepository.findByIds(bookDetails.genreIds);
+
+    if (!publishers.length || !authors.length || !genres.length) {
+      throw new NotFoundException('Publishers, Authors, or Genres not found');
+    }
+
     let bookToUpdate = await this.booksRepository
       .createQueryBuilder('book')
       .leftJoinAndSelect('book.authors', 'author')
@@ -98,6 +99,11 @@ export class BooksService {
       .leftJoinAndSelect('book.publishers', 'publisher')
       .where('book.id = :id', { id: id })
       .getOne();
+
+    if (!bookToUpdate) {
+      throw new NotFoundException('Book not found');
+    }
+
     bookToUpdate.bookTitle = bookDetails.bookTitle;
     bookToUpdate.copyWriteYear = bookDetails.copyWriteYear;
     bookToUpdate.editionNumber = bookDetails.editionNumber;
@@ -109,40 +115,54 @@ export class BooksService {
 
     return await this.booksRepository.save(bookToUpdate);
   }
+
   async findByGenre(subject: any) {
-    return this.booksRepository.find({ where: { subject: subject } });
+    return await this.booksRepository.find({ where: { subject: subject } });
   }
+
   async deleteBook(id: number) {
-    return this.booksRepository.delete(id);
+    const book = await this.booksRepository.findOneById(id);
+
+    if (!book) {
+      throw new NotFoundException('Book not found');
+    }
+
+    return await this.booksRepository.delete(id);
   }
+
   async myBooks(id: number) {
     return await this.booksRepository
       .createQueryBuilder('book')
       .select()
       .innerJoin(BookStock, 'book_stock', 'book.id = book_stock.bookId')
-      .innerJoin(
-        Reservation,
-        'reservation',
-        'book_stock.id = reservation.bookStockId',
-      )
+      .innerJoin(Reservation, 'reservation', 'book_stock.id = reservation.bookStockId')
       .where(
         "reservation.userId = :userId AND (reservation.reservationStatus = 'Active' OR reservation.reservationStatus ='Done') ",
         { userId: id },
       )
       .getMany();
   }
+
   async searchBooks(bookTitle: string) {
     return await this.booksRepository
       .createQueryBuilder('book')
       .where('bookTitle LIKE :title', { title: `%${bookTitle}%` })
       .getMany();
   }
+
   async getBookById(id: number) {
-    return await this.booksRepository.find({
+    const book = await this.booksRepository.findOne({
       where: { id: id },
       relations: ['genres', 'authors', 'publishers'],
     });
+
+    if (!book) {
+      throw new NotFoundException('Book not found');
+    }
+
+    return book;
   }
+
   async getBookStockCount(id: number) {
     return await this.bookStockRepo
       .createQueryBuilder()
@@ -150,15 +170,14 @@ export class BooksService {
       .where('bookId = :id', { id: id })
       .getCount();
   }
+
   async getMostBorrowedBooks() {
     return await this.booksRepository
       .createQueryBuilder()
       .select()
       .innerJoin(BookStock, 'book_stock', 'book.id = book_stock.bookId')
-      .innerJoin(
-        Reservation,
-        'reservation',
-        'book_stock.id = reservation.bookStockId').where('reservation.bookStockId').groupBy('book.id').getMany()
-      ;
+      .innerJoin(Reservation, 'reservation', 'book_stock.id = reservation.bookStockId')
+      .groupBy('book.id')
+      .getMany();
   }
 }
