+++ src/distributors/services/distributors.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { entityToLog } from 'src/books/mapper/logger.mapper';
 import { createDistributorDto } from 'src/distributors/dtos/createDistributor.to';
@@ -9,37 +9,52 @@ import { Repository } from 'typeorm';
 @Injectable()
 export class DistributorsService {
   constructor(
-    @InjectRepository(Distributor)
-    private distributorsRepository: Repository<Distributor>,
-    @InjectRepository(Audit) private loggerRepo: Repository<Audit>,
+    @InjectRepository(Distributor) private distributorsRepository: Repository<Distributor>,
+    @InjectRepository(Audit) private loggerRepository: Repository<Audit>,
   ) {}
+
   fetchDistributors() {
     return this.distributorsRepository.find();
   }
+
   async createDistributor(distributorDetails: createDistributorDto) {
     let newDistributor = this.distributorsRepository.create({
       ...distributorDetails,
     });
-    let newLog: Audit = entityToLog(
-      'New Distributor',
-      newDistributor,
-      'Distributors',
-    );
-    this.loggerRepo.save(newLog);
+
+    const newLog: Audit = entityToLog('New Distributor', newDistributor, 'Distributors');
+    await this.loggerRepository.save(newLog);
+
     return this.distributorsRepository.save(newDistributor);
   }
-  async updateDistributorById(
-    id: number,
-    distributorDetails: createDistributorDto,
-  ) {
-    return await this.distributorsRepository.update(id, {
-      ...distributorDetails,
-    });
+
+  async updateDistributorById(id: number, distributorDetails: createDistributorDto) {
+    const distributor = await this.distributorsRepository.findOneById(id);
+
+    if (!distributor) {
+      throw new NotFoundException('Distributor not found');
+    }
+
+    return await this.distributorsRepository.update(id, { ...distributorDetails });
   }
+
   async deleteDistributorById(id: number) {
+    const distributor = await this.distributorsRepository.findOneById(id);
+
+    if (!distributor) {
+      throw new NotFoundException('Distributor not found');
+    }
+
     return await this.distributorsRepository.delete(id);
   }
+
   async getDistributorById(id: number) {
-    return await this.distributorsRepository.findOneById(id);
+    const distributor = await this.distributorsRepository.findOneById(id);
+
+    if (!distributor) {
+      throw new NotFoundException('Distributor not found');
+    }
+
+    return distributor;
   }
 }
