+++ src/users/services/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Role } from 'src/typeorm/entities/Roles';
 import { User } from 'src/typeorm/entities/User';
@@ -10,13 +10,12 @@ import * as bcrypt from 'bcrypt';
 import { Audit } from 'src/typeorm/entities/Audit';
 import { entityToLog } from 'src/books/mapper/logger.mapper';
 
-
 @Injectable()
 export class UsersService {
   constructor(
     @InjectRepository(User) private usersRepository: Repository<User>,
     @InjectRepository(Role) private rolesRepository: Repository<Role>,
-    @InjectRepository(Audit) private loggerRepo: Repository<Audit>
+    @InjectRepository(Audit) private loggerRepo: Repository<Audit>,
   ) {}
 
   fetchUsers() {
@@ -24,23 +23,33 @@ export class UsersService {
   }
 
   async createUser(userDetails: createUserDto) {
-   // const roles = await this.rolesRepository.findByIds(userDetails.roleIds);
-    let newUser : User = dtoToEntity(userDetails);
-    //newUser.addRoles(roles);
-    newUser.password = await bcrypt.hash(newUser.password,10)
-    let newLog: Audit = entityToLog("New User",newUser,"Users")
-    this.loggerRepo.save(newLog)
+    let newUser: User = dtoToEntity(userDetails);
+    newUser.password = await bcrypt.hash(newUser.password, 10);
+    let newLog: Audit = entityToLog('New User', newUser, 'Users');
+    this.loggerRepo.save(newLog);
     return this.usersRepository.save(newUser);
   }
-  async findUserById(id: number){
+
+  async findUserById(id: number) {
     return await this.usersRepository.findOneById(id);
   }
+
   async updateUser(id: number, updateUserDetails: UpdateUserParams) {
-    return await this.usersRepository.update( id , { ...updateUserDetails });
+    const user = await this.usersRepository.findOneById(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    return await this.usersRepository.update(id, { ...updateUserDetails });
   }
+
   async deleteUser(id: number) {
+    const user = await this.usersRepository.findOneById(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
     return await this.usersRepository.delete(id);
   }
+
   async findOne(email: string): Promise<User | undefined> {
     return this.usersRepository.findOneBy({email});
   }
